Move NuGetSources into the inspection namespace since we will need it there
