+++ NuKeeper.Update/Process/DotNetUpdatePackageCommand.cs
@@ -1,9 +1,8 @@
-using System.Linq;
 using System.Threading.Tasks;
 using NuGet.Versioning;
-using NuKeeper.Inspection.Formats;
 using NuKeeper.Inspection.Logging;
 using NuKeeper.Inspection.RepositoryInspection;
+using NuKeeper.Inspection.Sources;
 using NuKeeper.Update.ProcessRunner;
 
 namespace NuKeeper.Update.Process
