+++ NuKeeper.Update/Process/NuGetFileRestoreCommand.cs
@@ -1,12 +1,10 @@
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using NuGet.Versioning;
-using NuKeeper.Inspection.Formats;
 using NuKeeper.Inspection.Logging;
 using NuKeeper.Inspection.RepositoryInspection;
+using NuKeeper.Inspection.Sources;
 using NuKeeper.Update.ProcessRunner;
 
 namespace NuKeeper.Update.Process
